+++ src/app/containers/AttributeMap.jsx
