+++ src/app/containers/CompetencyDetails.jsx
