+++ src/app/containers/ManageAttributes.jsx
