+++ src/app/containers/ResourceEdit.jsx
